Added tutorial, added world border, changed donation from Integer to Decimal, fixed school warp location

+++ plugin/src/main/java/net/wigoftime/open_komodo/commands/GenTip.java
@@ -28,7 +28,7 @@ public class GenTip extends Command {
 		}
 		
 		UUID uuid = UUID.fromString(args[0]);
-		int tip = Integer.parseInt(args[1]);
+		float tip = Float.parseFloat(args[1]);
 		
 		CustomPlayer.setDonated(uuid, tip);
